Add files via upload
Assignment (Data Engineering)
Question:
Our team works across verticals and is responsible for data, infrastructure, and analysis.
This data is from the IT industry where the client is dealing with services to customers after
purchasing a product. This means the client deals with data regarding tech support, warranties,
repairs, etc. For this assessment, we are interested in seeing some of your data skills and your
thought process around making sense of this kind of data.
Please automate the report such that it gets updated every month.

Data:
Zip file: GSPE_DE_dataset.zip

The objective of this assignment is to evaluate the candidates on the following skillsets:
Python coding / R coding
Making sense of the data
Thought process / presentation

The submission should include:
Python code / R code
Presentation on the steps followed
Summary of the final results
1.Please find the attached python code in mail:

2.Steps
* Retrieve the data
* Study and understand the dataset
*  Do the data cleaning
* Perform the EDA on that dataset
* Dig out the insight from the dataset

3.In Data cleaning I have deleted the duplicate rows, and replaced the null values with mean values, replaced the wrong spelled words in the correct order. and performed correlation operation and plotted a heat map of correlation data.

4.In Insight I have come plotted a few graphs which shows the Type and places of the customers,
* Most of the customers are from Hogwarts region.
* Most of the issues occurred with laptops than any other devices.
* People used to contact the customer service via voice is high.
* Though the count of hardware issue is high but There is no much difference in Hardwar and  Software issues.